requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
SQLAlchemy>=2.0
alembic>=1.13
PyMySQL>=1.1
python-dotenv>=1.0
pydantic-settings>=2.2
